% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build Network}
\usage{
buildNetwork(sites = sites, searchtype = "related", snowball = FALSE,
  nodes = NULL, edges = NULL, excludesites = NULL, delay = 2,
  maxurls = 10)
}
\arguments{
\item{sites}{a vector. Web sites for the Goolge related search.}

\item{searchtype}{a vector. Type of links search. 'related','links' or 'both' (default is 'related').}

\item{snowball}{a boolean. If TRUE snowball search results (default is FALSE).}

\item{nodes}{a data frame. nodes to append to new search to (default is NULL).}

\item{edges}{a data frame. edges to append to new search to (default is NULL).}

\item{excludesites}{(default is \code{NULL})}

\item{delay}{(default is 0)}

\item{maxurls}{(default is 10)}
}
\value{
What does this return?
}
\description{
Function for building a network of sites and links.
}
\details{
The buildNetwork function runs a Google 'related' search for the input sites
or parses the site HTML for hyperlinked sites.
}
\examples{
\dontrun{
buildNetwork(sites='["http://npr.org"]',searchtype = "related")
}
}
