% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hrefs.R
\name{get_hrefs}
\alias{get_hrefs}
\title{Get hrefs}
\usage{
get_hrefs(x, keep_regex = NULL, omit_regex = NULL,
  omit_bookmarks = TRUE, ...)
}
\arguments{
\item{x}{Either a character string of website of interest, or a
session has defined from the \pkg{rvest} function
\code{\link[rvest]{html_session}}}

\item{keep_regex}{a regular expression to be matched in the found hrefs, see
details.}

\item{omit_regex}{a regular expression to be matched in the found hrefs, see
details.}

\item{omit_bookmarks}{urls containing the "#" symbol will be omited from the
returned urls (Logical, defaults to \code{TRUE})}

\item{...}{not currently used}
}
\value{
A \code{sna_hrefs} object, which is a \code{data.frame} with the
following columns:
\describe{
\item{url}{<chr> the found urls, modified to be absolute urls}
\item{relative}{<logical> indicates whether or not the url relative to the domain of \code{x}}
}

The return object as additional attributes
\describe{
\item{session}{<session> the html session}
}

If the url or session does not resolve, the retruned \code{data.frame} will
have the aforementioned columns, but will have no rows.
}
\description{
Given a url, or html session, find the absolute urls of relative and external
links posted on the web page.
}
\details{
There are a few options for filtering the set of returned links:
\code{keep_regex}, \code{omit_regex}, and \code{omit_bookmarks}.  The first two
are regular expressions and will be applied to the set of links in order of
keep, then omit, that is: given a character vector of \code{links}, the use
of the \code{keep_regex} and \code{omit_regex} is equivalent to the following
two lines of code:

\verb{> links <- links[grepl(keep_regex, links)]}

\verb{> links <- links[!grepl(omit_regex, links)]}

Both \code{keep_regex} and \code{omit_regex} are optional.  You may consider
runing \code{get_hrefs} without filting results and inspect the returned
urls.  Post hoc filter would be viable, as would re-evaluating the
\code{get_hrefs} call with the wanted filters.

By default urls with the '#' symbol are omitted.  Set \code{omit_bookmarks =
FALSE} to include url with bookmarks in the return.
}
\examples{
\dontrun{
get_hrefs('neptuneinc.org')

## See the vignette for more details:
vignette("snaWeb", package = "snaWeb")
}

}
\seealso{
\code{vignette(topic = "snaWeb", package = "snaWeb")}
}
