% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_title.R
\name{get_page_title}
\alias{get_page_title}
\title{Get Page Title}
\usage{
get_page_title(site, timeout = 10, onTimeout = "error")
}
\arguments{
\item{site}{url as a character string, or a html session.}

\item{timeout}{A numeric specifying the maximum number of seconds the expression is allowed to run before being interrupted by the timeout.}

\item{onTimeout}{A character specifying what action to take if a timeout event occurs}
}
\description{
Get a page title
}
