% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphMetrics.R
\name{graphMetrics}
\alias{graphMetrics}
\title{Graph Metrics}
\usage{
graphMetrics(network)
}
\arguments{
\item{network}{list with nodes and nodes.}
}
\value{
What does this return?
}
\description{
Function for calculating grpah metrics.
}
\examples{
\dontrun{
graphMetrics(network)
}
}
