% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.sna_urls}
\alias{plot.sna_urls}
\title{Plotting Social Networks}
\usage{
\method{plot}{sna_urls}(x, method = "igraph", ...)
}
\arguments{
\item{x}{a \code{sna_urls} object}

\item{method}{the plotting method to use, defaults to basic R graphics via
\code{\link{igraph}{plot.igraph}}}

\item{...}{other arguments passed to plotting method}
}
\description{
Plotting tools for social network analysis
}
\details{
The possible plotting methods are, initially, all provided by the
\pkg{igraph} package.  The methods, noted below, can be modified via passing
arguments through \code{...}.

\describe{
\item{igraph}{uses \code{igraph:::plot.igraph}}
\item{tkplot}{uses \code{igraph::tkplot}}
\item{rglplot}{uses \code{igraph::rglplot}}
}
}
\seealso{
\code{\link{igraph}{plot.igraph}}
}
