% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_urls.R
\name{related_urls}
\alias{related_urls}
\title{Social Network Related Sites}
\usage{
related_urls(x, maxurls = 10L, delay = 2, excludesites = NULL, ...)
}
\arguments{
\item{x}{The url, as a character string, to base the google related site
serarch.}

\item{maxurls}{an integer value specifying the maximum number of related
sites to return.}

\item{delay}{minimum number of seconds to delay between sending queries to
Google.  The delay is needed to avoid having Google detect the queries as
coming from a bot.}

\item{excludesites}{(default is \code{NULL})}

\item{...}{passed to \code{\link[rvest]{html_session}}}
}
\value{
A \code{sna_sites} object.  This is a list with four elements:
\describe{
\item{nodes}{A \code{data.frame} describing the nodes of a graph
  \describe{
  \item{url}{url of the node}
  \item{is_root}{A character (looks like a logical, but is a character)
  with values "TRUE" and "FALSE" denoting if the node is a root (parent) node
  in the graph}
  \item{id}{node id}
  \item{name}{title of the website}
  \item{specification}{}
  }
}
\item{edges}{A \code{data.frame} describing the edges of a graph
  \describe{
  \item{name_from}{url of the parent (site) node for the edge}
  \item{name_to}{url of the child (site) note for the edge}
  \item{id}{edge id}
  \item{node_from}{node id of the parent node}
  \item{node_to}{node id of the child node}
  \item{rank}{The order the site was listed in the google search results.}
  \item{predicate}{}
  }
  }
\item{message}{a character string with the value "Blocked" or "Success."}
\item{is_blocked}{a logical of legnth 1}
}
}
\description{
Use Google's related:site advance search feature to fine websites related to
one or more sites of interest.
}
\details{
Performing searches on Google via a program can be difficult.  Google, and
many other search engines, monitor network traffic and will flag traffic that
might be coming from a bot.  Making several queries very quickly can result
in the queries being flag.  To address this issue we have a \code{delay}
argument to pause queries by a given number of seconds.

A preferable method to using the delay, and a method that *might* be
implemented at a later date, is to use the Google Custom Search JSON/Atom
API, \url{https://developers.google.com/custom-search/json-api/v1/overview}.
As of August 30, 2017, the "API provides 100 searches per day for free...
additional requests cost $5 per 1,000 queries up to 10k queries per day."

TODO: describe passing a timeout option via html_session, httr::config, ...
}
\examples{
\dontrun{
related_urls("neptuneinc.org")
}

}
