% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_url.R
\name{visit_url}
\alias{visit_url}
\title{Visit Url}
\usage{
visit_url(x, ...)
}
\arguments{
\item{x}{a url, specificied as a character string, or a html session}

\item{...}{not currently used.}
}
\value{
A \code{sna_visited_url} object.  This is a \code{data.frame} with
the following columns:
\describe{
\item{url}{the visited url}
\item{status}{the http status or error message status}
\item{access_date}{POSIXct date and time of the attempted visit}
\item{type}{content-type, if applicable}
\item{title}{title of the web page, if applicable}
}
}
\description{
Visit a Url and report social network analysis relevent details.
}
